don't rm if the directory is not exist

+++ lib/daitss/proc/wip/from_d1.rb
@@ -196,7 +196,7 @@ module Daitss
       end
 
     ensure
-      FileUtils.rm_r tdir
+      FileUtils.rm_r tdir if tdir and File.exist?(tdir)
     end
 
   end
